+++ lib/domain/doxygen.dart
@@ -1,0 +1,30 @@
+// Copyright (C) 2026 Vincent Hamp
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+/// Domain documentation
+///
+/// \file   domain/doxygen.dart
+/// \author Vincent Hamp
+/// \date   11/06/2026
+
+/// \page page_domain Domain
+/// \tableofcontents
+/// \todo Domain
+///
+/// <div class="section_buttons">
+/// | Previous     | Next           |
+/// | :----------- | -------------: |
+/// | \ref page_ui | \ref page_data |
+/// </div>