+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+##
+- Add VCC voltage and hardware revision to info screen ([#143](https://github.com/OpenRemise/Frontend/pull/143))
+- Blue power off icon indicates short circuit ([#143](https://github.com/OpenRemise/Frontend/pull/143))
+
 ## 0.7.1
 - Bugfix empty SSID is valid ([#133](https://github.com/OpenRemise/Frontend/issues/133))
 - Bugfix reload tries to clear caches ([#138](https://github.com/OpenRemise/Frontend/issues/138))