+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-##
+## 0.8.0
 - Add VCC voltage and hardware revision to info screen ([#143](https://github.com/OpenRemise/Frontend/pull/143))
 - Blue power off icon indicates short circuit ([#143](https://github.com/OpenRemise/Frontend/pull/143))